requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
bcrypt==4.1.3
httpx==0.27.0
python-multipart>=0.0.9
requests>=2.31.0
tzdata>=2024.2
stripe>=10.0.0
